// src/app/features/contabilidad/domain/models/cost-center-type.model.ts
export interface CostCenterType {
  id: number;
  name: string;
  code: string;
  description: string;
  createdAt: string;
  updatedAt: string;
  deletedAt: string | null;
}

export interface CostCenterTypePayload {
  name: string;
  code: string;
  description: string;
}

export interface CostCenterTypePage {
  items: CostCenterType[];
  links: {
    first: string | null;
    last: string | null;
    prev: string | null;
    next: string | null;
  };
  meta: {
    currentPage: number;
    from: number | null;
    lastPage: number;
    path: string;
    perPage: number;
    to: number | null;
    total: number;
  };
}